refactor(connection): split 1330-line mod.rs into submodules
Decomposes src/connection/mod.rs (1330 lines) into focused submodules.
Pure code-move: behavior byte-identical, only visibility/import glue added.

- error.rs      ConnectionError enum + to_runtime_error (51)
- auth.rs       get_auth_config + private get_user (215)
- elevation.rs  get_elevation_config (95)
- env.rs        setup_environment_ssh + setup_environment_local (133)
- session.rs    create_ssh_session + get_port_from_host +
                create_configured_ssh_session (SSH session assembly) (201)
- mod.rs        module doc + Connection enum/impl + create_connection factory +
                routing helpers (determine_connection_type/is_localhost/
                create_dummy_task) + re-exports (255)
- tests.rs      #[cfg(test)] unit tests, moved wholesale (416)

Visibility: setup_environment_local private->pub (called cross-module by
create_connection); get_user/get_port_from_host made pub inside their private
submodules so the test module can import them directly. External API intact
(crate::connection::{Connection, create_connection} + the pub helpers consumed
by komando.rs tests).
