+++ src/connection/mod.rs
@@ -45,71 +45,33 @@
 //! `context`). Verbose troubleshooting text was removed in favor of
 //! documentation; see `REFACTOR_PLAN.md` §2.2.
 
-use crate::defaults::Defaults;
-use crate::executor::CommandExecutor;
-use crate::local::LocalSession;
-use crate::models::ConnectionType;
-use crate::ssh::{Elevation, ElevationMethod, SSHAuthMethod, SSHSession};
-use crate::util::{host_display, task_display};
-use crate::validator::validate_host;
-use anyhow::Result;
-use mlua::{Error::RuntimeError, Lua, Table, Value};
-use secrecy::{ExposeSecret, SecretString};
-use std::env;
-use std::path::Path;
+mod auth;
+mod elevation;
+mod env;
+mod error;
+mod session;
 
 #[cfg(test)]
 pub mod test_utils;
 
-/// Errors raised while establishing an SSH or local connection.
-///
-/// The prior `troubleshooting` string fields, the `format_error` method, and
-/// the `get_*_troubleshooting` helpers were verbose boilerplate that cluttered
-/// every error message; they have been removed (see `REFACTOR_PLAN.md` §2.2).
-/// `Display` now renders only the structured fields. Detailed SSH debugging
-/// guidance belongs in documentation, not every error string.
-#[derive(Debug, Clone, thiserror::Error)]
-pub enum ConnectionError {
-    /// Host table failed validation (missing/invalid fields).
-    #[error("Host validation failed: {message} for host '{host}'")]
-    HostValidation { message: String, host: String },
+#[cfg(test)]
+mod tests;
 
-    /// SSH authentication failed.
-    #[error("SSH authentication failed: {message} for {user}@{host}")]
-    Authentication {
-        message: String,
-        host: String,
-        user: String,
-    },
+pub use auth::get_auth_config;
+pub use elevation::get_elevation_config;
+pub(crate) use env::setup_environment_local;
+pub use env::setup_environment_ssh;
+pub use error::ConnectionError;
+pub use session::{create_configured_ssh_session, create_ssh_session};
 
-    /// SSH TCP/session connection failed.
-    #[error("SSH connection failed: {message} for {host}:{port}")]
-    Connection {
-        message: String,
-        host: String,
-        port: u16,
-    },
-
-    /// SSH host-key verification failed.
-    #[error("SSH host key verification failed: {message} for host '{host}'")]
-    HostKeyVerification { message: String, host: String },
-
-    /// Connection-factory configuration error.
-    #[error("SSH configuration error: {message} in {context}")]
-    Configuration { message: String, context: String },
-}
-
-impl ConnectionError {
-    /// Convert to an `mlua::Error::RuntimeError` carrying the `Display` output.
-    ///
-    /// Kept as a named helper because ~30 call sites use the
-    /// `ConnectionError::Variant { ... }.to_runtime_error()` idiom at the mlua
-    /// boundary; renaming/chaining them all to `.into()` would be pure churn.
-    #[must_use]
-    pub fn to_runtime_error(self) -> mlua::Error {
-        RuntimeError(self.to_string())
-    }
-}
+use crate::executor::CommandExecutor;
+use crate::local::LocalSession;
+use crate::models::ConnectionType;
+use crate::ssh::SSHSession;
+use crate::util::host_display;
+use crate::validator::validate_host;
+use anyhow::Result;
+use mlua::{Lua, Table, Value};
 
 /// Unified connection interface that can represent either SSH or local connections
 #[derive(Clone, Debug)]
@@ -279,426 +241,6 @@ fn is_localhost(address: &str) -> bool {
     matches!(address, "localhost" | "127.0.0.1" | "::1")
 }
 
-/// Get the user for SSH connection from host, task, or defaults
-///
-/// This function extracts the user configuration using the same logic
-/// as the existing implementation in komando.rs.
-///
-/// # Arguments
-/// * `host` - Host configuration table
-/// * `task` - Task configuration table
-///
-/// # Returns
-/// * `mlua::Result<String>` - The resolved username
-///
-/// # Errors
-/// Returns an error if no user can be determined from any source
-fn get_user(host: &Table, task: &Table) -> mlua::Result<String> {
-    let defaults = Defaults::global();
-    let default_user = match defaults.user.read() {
-        Ok(user) => user.clone(),
-        Err(_) => {
-            return Err(RuntimeError(
-                "Failed to acquire read lock for default user".to_string(),
-            ));
-        }
-    };
-    let user = match host.get::<String>("user") {
-        Ok(user) => user,
-        Err(_) => match default_user {
-            Some(ref user) => user.clone(),
-            None => {
-                if let Ok(user) = env::var("USER") {
-                    user
-                } else {
-                    let task_display = task_display(task);
-                    let host_display = host_display(host);
-                    return Err(RuntimeError(format!(
-                        "No user specified for task '{task_display}' on host '{host_display}'. \
-                    Specify 'user' in host configuration, set default with komandan.defaults:set_user(), \
-                    or ensure USER environment variable is set."
-                    )));
-                }
-            }
-        },
-    };
-
-    Ok(user)
-}
-
-/// Get authentication configuration for SSH connections
-///
-/// This function extracts authentication method resolution logic from komando.rs
-/// and handles password, private key, and default key discovery.
-///
-/// # Arguments
-/// * `host` - Host configuration table
-/// * `task` - Task configuration table
-/// * `home_override` - Optional home directory override for testing
-///
-/// # Returns
-/// * `mlua::Result<(String, SSHAuthMethod)>` - Username and authentication method
-///
-/// # Errors
-/// Returns an error if:
-/// - No authentication method can be determined
-/// - Required environment variables are missing
-/// - Authentication configuration is invalid
-pub fn get_auth_config(
-    host: &Table,
-    task: &Table,
-    home_override: Option<&str>,
-) -> mlua::Result<(String, SSHAuthMethod)> {
-    let host_display = host_display(host);
-    let task_display = task_display(task);
-
-    let user = get_user(host, task).map_err(|e| {
-        ConnectionError::Configuration {
-            message: format!("Failed to determine user: {e}"),
-            context: format!("task '{task_display}' on host '{host_display}'"),
-        }
-        .to_runtime_error()
-    })?;
-
-    let defaults = Defaults::global();
-
-    let default_private_key_file = defaults
-        .private_key_file
-        .read()
-        .map_err(|_| {
-            ConnectionError::Configuration {
-                message: "Failed to read default private key file setting".to_string(),
-                context: "defaults access".to_string(),
-            }
-            .to_runtime_error()
-        })?
-        .clone();
-
-    let default_private_key_pass = defaults
-        .private_key_pass
-        .read()
-        .map_err(|_| {
-            ConnectionError::Configuration {
-                message: "Failed to read default private key passphrase setting".to_string(),
-                context: "defaults access".to_string(),
-            }
-            .to_runtime_error()
-        })?
-        .as_ref()
-        .map(|s: &SecretString| s.expose_secret().to_string());
-
-    let default_password = defaults
-        .password
-        .read()
-        .map_err(|_| {
-            ConnectionError::Configuration {
-                message: "Failed to read default password setting".to_string(),
-                context: "defaults access".to_string(),
-            }
-            .to_runtime_error()
-        })?
-        .as_ref()
-        .map(|s: &SecretString| s.expose_secret().to_string());
-
-    let ssh_auth_method = match host.get::<String>("private_key_file") {
-        Ok(private_key_file) => SSHAuthMethod::PublicKey {
-            private_key: private_key_file,
-            passphrase: host
-                .get::<String>("private_key_pass")
-                .ok()
-                .or(default_private_key_pass),
-        },
-        Err(_) => match default_private_key_file {
-            Some(ref private_key_file) => SSHAuthMethod::PublicKey {
-                private_key: private_key_file.clone(),
-                passphrase: host
-                    .get::<String>("private_key_pass")
-                    .ok()
-                    .or(default_private_key_pass),
-            },
-            None => match host.get::<String>("password") {
-                Ok(password) => SSHAuthMethod::Password(password),
-                Err(_) => {
-                    if let Some(ref password) = default_password {
-                        SSHAuthMethod::Password(password.clone())
-                    } else {
-                        // Check if SSH key auto-discovery is enabled
-                        if !*defaults.ssh_auto_discover_keys.read().map_err(|_| {
-                            ConnectionError::Configuration {
-                                message: "Failed to read ssh_auto_discover_keys setting"
-                                    .to_string(),
-                                context: "defaults access".to_string(),
-                            }
-                            .to_runtime_error()
-                        })? {
-                            return Err(ConnectionError::Authentication {
-                                message: "No authentication method available and SSH key auto-discovery is disabled".to_string(),
-                                host: host_display,
-                                user,
-                            }.to_runtime_error());
-                        }
-
-                        let home = if let Some(h) = home_override {
-                            h.to_string()
-                        } else {
-                            env::var("HOME").map_err(|_| {
-                                ConnectionError::Configuration {
-                                    message: "HOME environment variable not set".to_string(),
-                                    context: "SSH key discovery".to_string(),
-                                }
-                                .to_runtime_error()
-                            })?
-                        };
-                        let ed25519_path = format!("{home}/.ssh/id_ed25519");
-                        if Path::new(&ed25519_path).exists() {
-                            SSHAuthMethod::PublicKey {
-                                private_key: ed25519_path,
-                                passphrase: host
-                                    .get::<String>("private_key_pass")
-                                    .ok()
-                                    .or(default_private_key_pass),
-                            }
-                        } else {
-                            let rsa_path = format!("{home}/.ssh/id_rsa");
-                            if Path::new(&rsa_path).exists() {
-                                SSHAuthMethod::PublicKey {
-                                    private_key: rsa_path,
-                                    passphrase: host
-                                        .get::<String>("private_key_pass")
-                                        .ok()
-                                        .or(default_private_key_pass),
-                                }
-                            } else {
-                                return Err(ConnectionError::Authentication {
-                                    message: "No authentication method available".to_string(),
-                                    host: host_display,
-                                    user,
-                                }
-                                .to_runtime_error());
-                            }
-                        }
-                    }
-                }
-            },
-        },
-    };
-
-    Ok((user, ssh_auth_method))
-}
-
-/// Create and configure an SSH session with host key verification settings
-///
-/// This function extracts SSH session creation and configuration logic from komando.rs
-/// and handles host key verification and known hosts configuration.
-///
-/// # Arguments
-/// * `host` - Host configuration table
-///
-/// # Returns
-/// * `mlua::Result<SSHSession>` - A configured SSH session ready for connection
-///
-/// # Errors
-/// Returns an error if:
-/// - SSH session creation fails
-/// - Configuration parameters are invalid
-pub fn create_ssh_session(host: &Table) -> mlua::Result<SSHSession> {
-    let defaults = Defaults::global();
-    let mut ssh = SSHSession::new().map_err(|e| {
-        ConnectionError::Configuration {
-            message: format!("Failed to create SSH session: {e}"),
-            context: "SSH session initialization".to_string(),
-        }
-        .to_runtime_error()
-    })?;
-
-    let Ok(default_key_check) = defaults.key_check.read() else {
-        return Err(ConnectionError::Configuration {
-            message: "Failed to read default host key check setting".to_string(),
-            context: "defaults access".to_string(),
-        }
-        .to_runtime_error());
-    };
-
-    let host_key_check = host
-        .get::<Value>("host_key_check")
-        .map_or(true, |key_check| match key_check {
-            Value::Nil => *default_key_check,
-            Value::Boolean(false) => false,
-            _ => true,
-        });
-
-    let Ok(default_known_hosts_file) = defaults.known_hosts_file.read() else {
-        return Err(ConnectionError::Configuration {
-            message: "Failed to read default known hosts file setting".to_string(),
-            context: "defaults access".to_string(),
-        }
-        .to_runtime_error());
-    };
-
-    if host_key_check {
-        ssh.known_hosts_file = host
-            .get::<String>("known_hosts_file")
-            .map_or_else(|_| Some(default_known_hosts_file.clone()), Some);
-    }
-
-    Ok(ssh)
-}
-
-/// Get elevation configuration for privilege escalation
-///
-/// This function extracts privilege escalation configuration logic from komando.rs
-/// and handles sudo, su, and no elevation scenarios.
-///
-/// # Arguments
-/// * `host` - Host configuration table
-/// * `task` - Task configuration table
-///
-/// # Returns
-/// * `mlua::Result<Elevation>` - The elevation configuration
-///
-/// # Errors
-/// Returns an error if:
-/// - Default values cannot be read
-/// - Elevation method is invalid
-pub fn get_elevation_config(host: &Table, task: &Table) -> mlua::Result<Elevation> {
-    let defaults = Defaults::global();
-
-    let Ok(default_elevate) = defaults.elevate.read() else {
-        return Err(ConnectionError::Configuration {
-            message: "Failed to read default elevation setting".to_string(),
-            context: "defaults access".to_string(),
-        }
-        .to_runtime_error());
-    };
-
-    let task_elevate = task.get::<Value>("elevate")?;
-    let host_elevate = host.get::<Value>("elevate")?;
-
-    let elevate = if !task_elevate.is_nil() {
-        task_elevate.as_boolean().unwrap_or(false)
-    } else if !host_elevate.is_nil() {
-        host_elevate.as_boolean().unwrap_or(false)
-    } else {
-        *default_elevate
-    };
-
-    if !elevate {
-        return Ok(Elevation {
-            method: ElevationMethod::None,
-            as_user: None,
-        });
-    }
-
-    let Ok(default_elevation_method) = defaults.elevation_method.read() else {
-        return Err(ConnectionError::Configuration {
-            message: "Failed to read default elevation method setting".to_string(),
-            context: "defaults access".to_string(),
-        }
-        .to_runtime_error());
-    };
-
-    let elevation_method_str = task.get::<String>("elevation_method").unwrap_or_else(|_| {
-        host.get::<String>("elevation_method")
-            .unwrap_or_else(|_| default_elevation_method.clone())
-    });
-
-    let elevation_method = match elevation_method_str.as_str() {
-        "none" => Ok(ElevationMethod::None),
-        "sudo" => Ok(ElevationMethod::Sudo),
-        "su" => Ok(ElevationMethod::Su),
-        _ => Err(ConnectionError::Configuration {
-            message: format!("Unsupported elevation method: '{elevation_method_str}'"),
-            context: "elevation method configuration".to_string(),
-        }
-        .to_runtime_error()),
-    };
-
-    let default_as_user = match defaults.as_user.read() {
-        Ok(as_user) => as_user.clone(),
-        Err(_) => {
-            return Err(ConnectionError::Configuration {
-                message: "Failed to read default as_user setting".to_string(),
-                context: "defaults access".to_string(),
-            }
-            .to_runtime_error());
-        }
-    };
-
-    let as_user = task.get::<Option<String>>("as_user").unwrap_or_else(|_| {
-        host.get::<Option<String>>("as_user")
-            .unwrap_or(default_as_user)
-    });
-
-    Ok(Elevation {
-        method: elevation_method?,
-        as_user,
-    })
-}
-
-/// Set up environment variables for SSH sessions
-///
-/// This function extracts environment variable setup logic from komando.rs
-/// and handles defaults, host-level, and task-level environment variables.
-///
-/// # Arguments
-/// * `ssh` - Mutable reference to SSH session
-/// * `host` - Host configuration table
-/// * `task` - Task configuration table
-///
-/// # Returns
-/// * `mlua::Result<()>` - Success or error
-///
-/// # Errors
-/// Returns an error if:
-/// - Default values cannot be read
-/// - Environment variable tables cannot be processed
-pub fn setup_environment_ssh(ssh: &mut SSHSession, host: &Table, task: &Table) -> mlua::Result<()> {
-    let defaults = Defaults::global();
-
-    let Ok(default_env) = defaults.env.read() else {
-        return Err(ConnectionError::Configuration {
-            message: "Failed to read default environment variables".to_string(),
-            context: "defaults access".to_string(),
-        }
-        .to_runtime_error());
-    };
-
-    let env_host = host.get::<Option<Table>>("env")?;
-    let env_task = task.get::<Option<Table>>("env")?;
-
-    for (key, value) in default_env.iter() {
-        ssh.set_env(key, value);
-    }
-
-    if let Some(env_host) = env_host {
-        for pair in env_host.pairs::<String, String>() {
-            let (key, value) = pair.map_err(|e| {
-                ConnectionError::Configuration {
-                    message: format!("Invalid host environment variable: {e}"),
-                    context: "host environment variable processing".to_string(),
-                }
-                .to_runtime_error()
-            })?;
-            ssh.set_env(&key, &value);
-        }
-    }
-
-    if let Some(env_task) = env_task {
-        for pair in env_task.pairs::<String, String>() {
-            let (key, value) = pair.map_err(|e| {
-                ConnectionError::Configuration {
-                    message: format!("Invalid task environment variable: {e}"),
-                    context: "task environment variable processing".to_string(),
-                }
-                .to_runtime_error()
-            })?;
-            ssh.set_env(&key, &value);
-        }
-    }
-
-    Ok(())
-}
-
 /// Create a minimal task table for functions that don't have task context
 ///
 /// This allows reuse of existing functions that expect both host and task parameters.
@@ -710,621 +252,4 @@ pub fn setup_environment_ssh(ssh: &mut SSHSession, host: &Table, task: &Table) -
 /// * `mlua::Result<Table>` - An empty task table
 fn create_dummy_task(lua: &Lua) -> mlua::Result<Table> {
     lua.create_table()
-}
-
-/// Set up environment variables for local sessions
-///
-/// This function applies environment variables to local sessions using the same
-/// logic as SSH sessions for consistency.
-///
-/// # Arguments
-/// * `local` - Mutable reference to local session
-/// * `host` - Host configuration table
-/// * `task` - Task configuration table
-///
-/// # Returns
-/// * `mlua::Result<()>` - Success or error
-fn setup_environment_local(
-    local: &mut LocalSession,
-    host: &Table,
-    task: &Table,
-) -> mlua::Result<()> {
-    let defaults = Defaults::global();
-
-    let Ok(default_env) = defaults.env.read() else {
-        return Err(ConnectionError::Configuration {
-            message: "Failed to read default environment variables".to_string(),
-            context: "defaults access".to_string(),
-        }
-        .to_runtime_error());
-    };
-
-    let env_host = host.get::<Option<Table>>("env")?;
-    let env_task = task.get::<Option<Table>>("env")?;
-
-    for (key, value) in default_env.iter() {
-        local.set_env(key, value);
-    }
-
-    if let Some(env_host) = env_host {
-        for pair in env_host.pairs::<String, String>() {
-            let (key, value) = pair.map_err(|e| {
-                ConnectionError::Configuration {
-                    message: format!("Invalid host environment variable: {e}"),
-                    context: "host environment variable processing".to_string(),
-                }
-                .to_runtime_error()
-            })?;
-            local.set_env(&key, &value);
-        }
-    }
-
-    if let Some(env_task) = env_task {
-        for pair in env_task.pairs::<String, String>() {
-            let (key, value) = pair.map_err(|e| {
-                ConnectionError::Configuration {
-                    message: format!("Invalid task environment variable: {e}"),
-                    context: "task environment variable processing".to_string(),
-                }
-                .to_runtime_error()
-            })?;
-            local.set_env(&key, &value);
-        }
-    }
-
-    Ok(())
-}
-
-/// Get the port for SSH connection from host or defaults
-///
-/// # Arguments
-/// * `host` - Host configuration table
-///
-/// # Returns
-/// * `mlua::Result<u16>` - The resolved port number
-fn get_port_from_host(host: &Table) -> mlua::Result<u16> {
-    let defaults = Defaults::global();
-
-    // Try to get port from host configuration first
-    if let Ok(port) = host.get::<u16>("port") {
-        Ok(port)
-    } else {
-        // Fall back to default port
-        let Ok(default_port) = defaults.port.read() else {
-            return Err(ConnectionError::Configuration {
-                message: "Failed to read default port setting".to_string(),
-                context: "defaults access".to_string(),
-            }
-            .to_runtime_error());
-        };
-        Ok(*default_port)
-    }
-}
-
-/// Create a fully configured SSH session using existing komando.rs logic
-///
-/// This function combines authentication, session creation, and configuration setup
-/// to provide a ready-to-use SSH session.
-///
-/// # Arguments
-/// * `host_table` - Host configuration table
-/// * `task` - Task configuration table
-///
-/// # Returns
-/// * `mlua::Result<SSHSession>` - A fully configured SSH session
-///
-/// # Errors
-/// Returns an error if:
-/// - Authentication configuration fails
-/// - SSH session creation fails
-/// - Connection establishment fails
-/// - Configuration setup fails
-pub fn create_configured_ssh_session(host_table: &Table, task: &Table) -> mlua::Result<SSHSession> {
-    let host_display = host_display(host_table);
-
-    // Use existing authentication configuration logic
-    let (user, auth_method) = get_auth_config(host_table, task, None)?;
-
-    // Use existing SSH session creation logic
-    let mut ssh = create_ssh_session(host_table).map_err(|e| {
-        ConnectionError::Configuration {
-            message: format!("Failed to create SSH session: {e}"),
-            context: format!("SSH session creation for host '{host_display}'"),
-        }
-        .to_runtime_error()
-    })?;
-
-    // Extract connection parameters
-    let address = host_table.get::<String>("address").map_err(|e| {
-        ConnectionError::Configuration {
-            message: format!("Missing or invalid address: {e}"),
-            context: format!("host '{host_display}'"),
-        }
-        .to_runtime_error()
-    })?;
-
-    let port = get_port_from_host(host_table).map_err(|e| {
-        ConnectionError::Configuration {
-            message: format!("Failed to get port: {e}"),
-            context: format!("host '{host_display}'"),
-        }
-        .to_runtime_error()
-    })?;
-
-    // Connect using existing logic with error type classification
-    ssh.connect(&address, port, &user, auth_method)
-        .map_err(|e| {
-            let error_msg = e.to_string();
-
-            // Classify the underlying error by its message content
-            let is_auth = error_msg.contains("authentication")
-                || error_msg.contains("Authentication")
-                || error_msg.contains("auth")
-                || error_msg.contains("login")
-                || error_msg.contains("password")
-                || error_msg.contains("key")
-                || error_msg.contains("Permission denied");
-            let is_host_key = error_msg.contains("host key")
-                || error_msg.contains("Host key")
-                || error_msg.contains("known_hosts")
-                || error_msg.contains("verification");
-
-            if is_auth {
-                ConnectionError::Authentication {
-                    message: error_msg,
-                    host: format!("{address}:{port}"),
-                    user: user.clone(),
-                }
-                .to_runtime_error()
-            } else if is_host_key {
-                ConnectionError::HostKeyVerification {
-                    message: error_msg,
-                    host: address.clone(),
-                }
-                .to_runtime_error()
-            } else {
-                ConnectionError::Connection {
-                    message: error_msg,
-                    host: address.clone(),
-                    port,
-                }
-                .to_runtime_error()
-            }
-        })?;
-
-    // Apply elevation and environment configuration
-    ssh.elevation = get_elevation_config(host_table, task).map_err(|e| {
-        ConnectionError::Configuration {
-            message: format!("Failed to configure elevation: {e}"),
-            context: format!("elevation setup for host '{host_display}'"),
-        }
-        .to_runtime_error()
-    })?;
-
-    setup_environment_ssh(&mut ssh, host_table, task).map_err(|e| {
-        ConnectionError::Configuration {
-            message: format!("Failed to setup SSH environment: {e}"),
-            context: format!("environment setup for host '{host_display}'"),
-        }
-        .to_runtime_error()
-    })?;
-
-    Ok(ssh)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::create_lua;
-
-    #[test]
-    fn test_create_connection_local() -> mlua::Result<()> {
-        let lua = create_lua()?;
-        let host_table = lua.create_table()?;
-        host_table.set("address", "localhost")?;
-
-        let connection = create_connection(&lua, &Value::Table(host_table))?;
-
-        match connection {
-            Connection::Local(_) => {}
-            Connection::SSH(_) => panic!("Expected local connection for localhost"),
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_create_connection_ssh_factory_logic() -> mlua::Result<()> {
-        let lua = create_lua()?;
-        let host_table = lua.create_table()?;
-        host_table.set("address", "remote.example.com")?;
-        host_table.set("user", "testuser")?;
-        host_table.set("password", "testpass")?;
-
-        // Test that the factory correctly identifies this as SSH
-        let connection_type = determine_connection_type(&host_table)?;
-        assert_eq!(connection_type, ConnectionType::SSH);
-
-        // Test that we can create the dummy task
-        let task = create_dummy_task(&lua)?;
-        assert_eq!(task.len()?, 0);
-
-        // Test that we can get auth config
-        let (user, auth) = get_auth_config(&host_table, &task, None)?;
-        assert_eq!(user, "testuser");
-        match auth {
-            SSHAuthMethod::Password(pass) => assert_eq!(pass, "testpass"),
-            SSHAuthMethod::PublicKey { .. } => panic!("Expected Password authentication"),
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_create_connection_explicit_local() -> mlua::Result<()> {
-        let lua = create_lua()?;
-        let host_table = lua.create_table()?;
-        host_table.set("address", "remote.example.com")?;
-        host_table.set("connection", "local")?;
-
-        let connection = create_connection(&lua, &Value::Table(host_table))?;
-
-        match connection {
-            Connection::Local(_) => {}
-            Connection::SSH(_) => panic!("Expected local connection when explicitly set"),
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_create_connection_explicit_ssh_factory_logic() -> mlua::Result<()> {
-        let lua = create_lua()?;
-        let host_table = lua.create_table()?;
-        host_table.set("address", "localhost")?;
-        host_table.set("connection", "ssh")?;
-        host_table.set("user", "testuser")?;
-        host_table.set("password", "testpass")?;
-
-        // Test that the factory correctly identifies this as SSH even for localhost
-        let connection_type = determine_connection_type(&host_table)?;
-        assert_eq!(connection_type, ConnectionType::SSH);
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_create_connection_with_environment() -> mlua::Result<()> {
-        let lua = create_lua()?;
-        let host_table = lua.create_table()?;
-        host_table.set("address", "localhost")?;
-
-        let env_table = lua.create_table()?;
-        env_table.set("TEST_VAR", "test_value")?;
-        host_table.set("env", env_table)?;
-
-        let connection = create_connection(&lua, &Value::Table(host_table))?;
-
-        match connection {
-            Connection::Local(_) => {}
-            Connection::SSH(_) => panic!("Expected local connection for localhost"),
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_create_dummy_task() -> mlua::Result<()> {
-        let lua = create_lua()?;
-        let task = create_dummy_task(&lua)?;
-
-        // Should be an empty table
-        assert_eq!(task.len()?, 0);
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_get_port_from_host() -> mlua::Result<()> {
-        let lua = create_lua()?;
-
-        // Test with explicit port
-        let host_table = lua.create_table()?;
-        host_table.set("port", 2222)?;
-        let port = get_port_from_host(&host_table)?;
-        assert_eq!(port, 2222);
-
-        // Test with default port - reset defaults first
-        lua.load(mlua::chunk! {
-            komandan.defaults:set_port(22)
-        })
-        .exec()?;
-
-        let host_table = lua.create_table()?;
-        let port = get_port_from_host(&host_table)?;
-        assert_eq!(port, 22); // Should use default
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_determine_connection_type_localhost() -> mlua::Result<()> {
-        let lua = create_lua()?;
-        let host_table = lua.create_table()?;
-        host_table.set("address", "localhost")?;
-
-        let conn_type = determine_connection_type(&host_table)?;
-        assert_eq!(conn_type, ConnectionType::Local);
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_determine_connection_type_127_0_0_1() -> mlua::Result<()> {
-        let lua = create_lua()?;
-        let host_table = lua.create_table()?;
-        host_table.set("address", "127.0.0.1")?;
-
-        let conn_type = determine_connection_type(&host_table)?;
-        assert_eq!(conn_type, ConnectionType::Local);
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_determine_connection_type_ipv6_localhost() -> mlua::Result<()> {
-        let lua = create_lua()?;
-        let host_table = lua.create_table()?;
-        host_table.set("address", "::1")?;
-
-        let conn_type = determine_connection_type(&host_table)?;
-        assert_eq!(conn_type, ConnectionType::Local);
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_determine_connection_type_remote() -> mlua::Result<()> {
-        let lua = create_lua()?;
-        let host_table = lua.create_table()?;
-        host_table.set("address", "remote.example.com")?;
-
-        let conn_type = determine_connection_type(&host_table)?;
-        assert_eq!(conn_type, ConnectionType::SSH);
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_determine_connection_type_explicit() -> mlua::Result<()> {
-        let lua = create_lua()?;
-        let host_table = lua.create_table()?;
-        host_table.set("address", "localhost")?;
-        host_table.set("connection", "ssh")?;
-
-        let conn_type = determine_connection_type(&host_table)?;
-        assert_eq!(conn_type, ConnectionType::SSH);
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_is_localhost() {
-        assert!(is_localhost("localhost"));
-        assert!(is_localhost("127.0.0.1"));
-        assert!(is_localhost("::1"));
-        assert!(!is_localhost("remote.example.com"));
-        assert!(!is_localhost("192.168.1.1"));
-    }
-
-    #[test]
-    fn test_connection_type() -> mlua::Result<()> {
-        let lua = create_lua()?;
-
-        // Test local connection type
-        let local_host = lua.create_table()?;
-        local_host.set("address", "localhost")?;
-        let local_conn = create_connection(&lua, &Value::Table(local_host))?;
-        assert_eq!(local_conn.connection_type(), ConnectionType::Local);
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_get_auth_config() -> anyhow::Result<()> {
-        let lua = create_lua()?;
-        let host = lua.create_table()?;
-
-        // Test with user in host
-        host.set("address", "localhost")?;
-        host.set("user", "testuser")?;
-        host.set("private_key_file", "/path/to/key")?;
-
-        let module_params = lua.create_table()?;
-        module_params.set("cmd", "echo test")?;
-        let module = lua
-            .load(mlua::chunk! {
-                return komandan.modules.cmd($module_params)
-            })
-            .eval::<Table>()?;
-        let task = lua.create_table()?;
-        task.set(1, module)?;
-
-        let (user, auth) = get_auth_config(&host, &task, None)?;
-        assert_eq!(user, "testuser");
-        match auth {
-            SSHAuthMethod::PublicKey {
-                private_key,
-                passphrase,
-            } => {
-                assert_eq!(private_key, "/path/to/key");
-                assert!(passphrase.is_none());
-            }
-            SSHAuthMethod::Password(_) => panic!("Expected PublicKey authentication"),
-        }
-
-        // Test with password auth
-        host.set("private_key_file", Value::Nil)?;
-        host.set("password", "testpass")?;
-        let (_, auth) = get_auth_config(&host, &task, None)?;
-        match auth {
-            SSHAuthMethod::Password(pass) => assert_eq!(pass, "testpass"),
-            SSHAuthMethod::PublicKey { .. } => panic!("Expected Password authentication"),
-        }
-
-        // Test with no authentication method
-        host.set("password", Value::Nil)?;
-        let temp_dir =
-            tempfile::tempdir().map_err(|e| anyhow::anyhow!("failed to create temp dir: {e}"))?;
-        let home_path = temp_dir.path().display().to_string();
-        let result = get_auth_config(&host, &task, Some(&home_path));
-        assert!(result.is_err());
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_get_user() -> mlua::Result<()> {
-        let lua = create_lua()?;
-        let host = lua.create_table()?;
-        let task = lua.create_table()?;
-
-        // Test with user in host
-        host.set("user", "testuser")?;
-        let user = get_user(&host, &task)?;
-        assert_eq!(user, "testuser");
-
-        // Test with no user specified (should fall back to environment)
-        host.set("user", Value::Nil)?;
-        let user = get_user(&host, &task);
-        // This should either succeed with the current USER env var or fail
-        // We can't predict the exact behavior since it depends on the environment
-        assert!(user.is_ok() || user.is_err());
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_create_ssh_session() -> mlua::Result<()> {
-        let lua = create_lua()?;
-        let host = lua.create_table()?;
-        host.set("address", "localhost")?;
-
-        // Test with default settings
-        let ssh = create_ssh_session(&host)?;
-        assert!(ssh.known_hosts_file.is_some());
-
-        // Test with host key check disabled
-        host.set("host_key_check", false)?;
-        let ssh = create_ssh_session(&host)?;
-        assert!(ssh.known_hosts_file.is_none());
-
-        // Test with custom known_hosts file
-        host.set("known_hosts_file", "/path/to/known_hosts")?;
-        host.set("host_key_check", true)?;
-        let ssh = create_ssh_session(&host)?;
-        assert_eq!(
-            ssh.known_hosts_file,
-            Some("/path/to/known_hosts".to_string())
-        );
-
-        // Test with known_hosts from defaults
-        host.set("known_hosts_file", Value::Nil)?;
-        lua.load(mlua::chunk! {
-            komandan.defaults:set_known_hosts_file("/default/known_hosts")
-        })
-        .exec()?;
-        let ssh = create_ssh_session(&host)?;
-        assert_eq!(
-            ssh.known_hosts_file,
-            Some("/default/known_hosts".to_string())
-        );
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_get_elevation_config() -> mlua::Result<()> {
-        let lua = create_lua()?;
-        let host = lua.create_table()?;
-        let task = lua.create_table()?;
-
-        // Test with no elevation
-        let elevation = get_elevation_config(&host, &task)?;
-        assert!(matches!(
-            elevation,
-            Elevation {
-                method: ElevationMethod::None,
-                as_user: None
-            }
-        ));
-
-        // Test with elevation from task
-        task.set("elevate", true)?;
-        let elevation = get_elevation_config(&host, &task)?;
-        assert!(matches!(
-            elevation,
-            Elevation {
-                method: ElevationMethod::Sudo,
-                as_user: None
-            }
-        ));
-
-        // Test with custom elevation method
-        task.set("elevation_method", "su")?;
-        let elevation = get_elevation_config(&host, &task)?;
-        assert!(matches!(
-            elevation,
-            Elevation {
-                method: ElevationMethod::Su,
-                as_user: None
-            }
-        ));
-
-        // Test invalid elevation method
-        task.set("elevation_method", "invalid")?;
-        assert!(get_elevation_config(&host, &task).is_err());
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_setup_environment_ssh() -> mlua::Result<()> {
-        let lua = create_lua()?;
-        let mut ssh = SSHSession::new()
-            .map_err(|e| RuntimeError(format!("Failed to create SSH session: {e}")))?;
-        let host = lua.create_table()?;
-        let task = lua.create_table()?;
-
-        // Test with environment variables at all levels
-        let env_host = lua.create_table()?;
-        env_host.set("HOST_VAR", "host_value")?;
-        env_host.set("OVERRIDE_VAR", "host_override")?; // This should be overridden by task
-        host.set("env", env_host)?;
-
-        let env_task = lua.create_table()?;
-        env_task.set("TASK_VAR", "task_value")?;
-        env_task.set("OVERRIDE_VAR", "task_override")?; // This should override host value
-        task.set("env", env_task)?;
-
-        setup_environment_ssh(&mut ssh, &host, &task)?;
-
-        // We can't directly test the environment variables since SSHSession doesn't expose them
-        // But we can verify the function completes without error
-        Ok(())
-    }
-
-    #[test]
-    fn test_setup_environment_ssh_empty() -> mlua::Result<()> {
-        let lua = create_lua()?;
-        let mut ssh = SSHSession::new()
-            .map_err(|e| RuntimeError(format!("Failed to create SSH session: {e}")))?;
-        let host = lua.create_table()?;
-        let task = lua.create_table()?;
-
-        // Test with no environment variables
-        setup_environment_ssh(&mut ssh, &host, &task)?;
-
-        Ok(())
-    }
 }
